Clean out old reports
When we renamed the workflows in the SDK repo, our worklfow
run numbering got reset. This messed up the logic in the
allure-report-action that removes old runs: because the newer
runs have a lower number, they were removed instead of the
older runs, which had a greater number. Manually clearing
out the old runs will restore everything to normal working
order.

Signed-off-by: Sam Friedman <sam@golioth.io>